Decouple constants from NumberGenerator.
NG now takes min & max values on init.

+++ src/constants.rs
@@ -1,3 +1,3 @@
 pub const INVALID_CHOICE: &str = "Invalid choice!";
-pub const MIN_SECRET: u8 = 0;
-pub const MAX_SECRET: u8 = 100;
+pub const MIN_SECRET: usize = 0;
+pub const MAX_SECRET: usize = 100;
